0915 django - signin

+++ Django/0915/04_django_auth/articles/views.py
@@ -14,6 +14,7 @@ def index(request):
     }
     return render(request, 'articles/index.html', context)
 
+
 @login_required
 @require_http_methods(['GET', 'POST'])
 def create(request):
@@ -38,12 +39,15 @@ def detail(request, pk):
     }
     return render(request, 'articles/detail.html', context)
 
-@login_required
+
+# @login_required
 @require_POST
 def delete(request, pk):
-    article = get_object_or_404(Article, pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        article.delete()
     return redirect('articles:index')
+
 
 @login_required
 @require_http_methods(['GET', 'POST'])
